Updates code to 2022.03 version of FMS, FMScoupler
Updates namelists to allow for backward reproducibility
Adds ncrc5.intel-classic platform

NB : Some of these XMLs may not be fully "fit for purpose".
Check if the XML has a ncrc5 platform.
In particular, some c384 XMLs may have an issue reproducing old answers
due to a change in the FV3 code which may (or may not) be fully tested
in these XMLs.
Further testing will recover old answers and be checked in if necessary.